requirements.txt
langchain>=0.1.0
langgraph>=0.0.26
qdrant-client>=1.7.0
streamlit>=1.31.0
PyMuPDF>=1.23.0
sentence-transformers>=2.3.0
python-dotenv>=1.0.1
pydantic>=2.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.19
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=4.0.0
